use can download inputfile and outputfile

+++ myproject/keepcodeapp/views.py
@@ -11,44 +11,76 @@ def program_file(request):
     username = request.session['username']
     fs = FileSystemStorage()
     visited = {}
+    inputfile = {}
+    outputfile = {}
     message = []
+    fname = "C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\"
     for i in range(1,51):
         message.append(str(i))
-        filename = "C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\"+username+"_program"+str(i)+".c"
+        filename = fname + username + "_program" + str(i) + ".c"
         if fs.exists(filename):
             visited[str(i)] = str(1)
+        filename = fname + "input_program" + str(i) + ".txt"
+        if fs.exists(filename):
+            inputfile[str(i)] = str(1)
+        filename = fname + "output_program" + str(i) + ".txt"
+        if fs.exists(filename):
+            outputfile[str(i)] = str(1)
     c = {}
     c['message'] = message
     c['visited'] = visited
-    c["one"] = "1"
+    c['inputfile'] = inputfile
+    c['outputfile'] = outputfile
     return render(request,'index.html',c)
 
 @login_required(login_url="/loginmodule/login")
 def upload_file(request):
-    id = request.GET.get('id')
-    username = request.session['username']
-    file1 = request.FILES["codefile"]
-    fs = FileSystemStorage()
-    filename = "C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\"+username+"_program"+str(id)+".c"
-    if fs.exists(filename):
-        fs.delete(filename)
-    file2 = fs.save("C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\"+username+"_program"+str(id)+".c", file1)
-    message = []
-    visited = {}
-    for i in range(1,51):
-        message.append(str(i))
-        filename = "C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\"+username+"_program"+str(i)+".c"
+    try:
+        id = request.GET.get('id')
+        username = request.session['username']
+        file1 = request.FILES["codefile"]
+        fs = FileSystemStorage()
+        filename = "C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\"+username+"_program"+str(id)+".c"
         if fs.exists(filename):
-            visited[str(i)] = str(1)
-    c = {}
-    c['message'] = message
-    c['visited'] = visited
-    c["one"] = "1"
-    return render(request,'index.html',c)
+            fs.delete(filename)
+        file2 = fs.save("C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\"+username+"_program"+str(id)+".c", file1)
+        return HttpResponseRedirect('/keepcodeap/program_file/')
+    except:
+        return HttpResponseRedirect('/keepcodeap/program_file/')
 
 @login_required(login_url="/loginmodule/login")
 def upload_input(request):
-    return render(request,'index.html',c)
+    try:
+        if not request.user.is_superuser:
+            return HttpResponseRedirect('/admin/')
+        else:
+            id = request.GET.get('id')
+            file1 = request.FILES["inputfile"]
+            fs = FileSystemStorage()
+            filename = "C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\input_program"+str(id)+".txt"
+            if fs.exists(filename):
+                fs.delete(filename)
+            file2 = fs.save("C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\input_program"+str(id)+".txt", file1)
+            return HttpResponseRedirect('/keepcodeap/program_file/')
+    except:
+        return HttpResponseRedirect('/keepcodeap/program_file/')
+
+@login_required(login_url="/loginmodule/login")
+def upload_output(request):
+    try:
+        if not request.user.is_superuser:
+            return HttpResponseRedirect('/admin/')
+        else:
+            id = request.GET.get('id')
+            file1 = request.FILES["outputfile"]
+            fs = FileSystemStorage()
+            filename = "C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\output_program"+str(id)+".txt"
+            if fs.exists(filename):
+                fs.delete(filename)
+            file2 = fs.save("C:\\Users\\Naimish\\Desktop\\Project1\\myproject\\keepcodeapp\\static\\output_program"+str(id)+".txt", file1)
+            return HttpResponseRedirect('/keepcodeap/program_file/')
+    except:
+        return HttpResponseRedirect('/keepcodeap/program_file/')
 
 @login_required(login_url="/loginmodule/login")
 def add_user(request):
